Drop unused types reported by rustc 1.90.
In one case the type was only used inside a test, so I moved it in there.
